corrected error in checksum calculation winternitz
Corrected an error in the winternitz checksum-calculation: When calculating C,
the parts of m (m_0,...,m_t1-1) are used. Before, the sum of C went from 0 to t.
That doesn't make sense, as there are only t1-1 parts of m. So the top border of the sum was corrected to t1-1.

Also inserted subsection structure for xmss.